requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.2.1
pika==1.3.2
#paddlepaddle==3.2.1 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddleocr==3.2.0
Pillow==10.1.0
PyMuPDF==1.21.0